requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.2
python-multipart==0.0.6
google-generativeai==0.3.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
httpx==0.25.0